:sparkles: feat: 필터링 조회 기능 개선
str 쿼리 추가
