+++ src/utils/index.js
@@ -4,3 +4,4 @@ export * from "./errorCodes";
 export * from "./multer";
 export * from "./sorting-cart";
 export * from "./pagination";
+export * from "./make-filterObj";
